// src/app/http/model/coin.ts
/**
 * VMCS
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface Coin {
  id: number;
  /**
   * ex. $1, 50 c
   */
  name: Coin.NameEnum;
  /**
   * face value ($1 for 100)
   */
  value: Coin.ValueEnum;
  /**
   * stock
   */
  quantity: number;
  weight: number;
}
export namespace Coin {
  export type NameEnum = '5c' | '10c' | '20c' | '50c' | '$1';
  export const NameEnum = {
    _5c: '5c' as NameEnum,
    _10c: '10c' as NameEnum,
    _20c: '20c' as NameEnum,
    _50c: '50c' as NameEnum,
    _1: '$1' as NameEnum,
  };
  export type ValueEnum = 5 | 10 | 20 | 50 | 100;
  export const ValueEnum = {
    NUMBER_5: 5 as ValueEnum,
    NUMBER_10: 10 as ValueEnum,
    NUMBER_20: 20 as ValueEnum,
    NUMBER_50: 50 as ValueEnum,
    NUMBER_100: 100 as ValueEnum,
  };
}